requirements.txt
streamlit
scikit-learn
opencv-python
scikit-image
numpy
pandas
matplotlib
imbalanced-learn
joblib
Pillow
opencv-python-headless